Reference

Contents

Reference

Contents

Index

Types

Methods

Convenience Constructors

Index

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.GaussianMixtureModel

GenerativeModelProtocols.GaussianMixtureModel

GenerativeModelProtocols.GenerativeAdversarialNetwork

GenerativeModelProtocols.GenerativeAdversarialNetwork

GenerativeModelProtocols.GenerativeModelProtocol

GenerativeModelProtocols.GenerativeModelProtocol

GenerativeModelProtocols.NormalizingFlow

GenerativeModelProtocols.NormalizingFlow

GenerativeModelProtocols.TabularDenoiser

GenerativeModelProtocols.TabularDenoiser

GenerativeModelProtocols.VariationalAutoencoder

GenerativeModelProtocols.VariationalAutoencoder

GenerativeModelProtocols.VariationalAutoencoder

GenerativeModelProtocols.VariationalAutoencoder

GenerativeModelProtocols.categorize

GenerativeModelProtocols.categorize

GenerativeModelProtocols.decode

GenerativeModelProtocols.decode

GenerativeModelProtocols.default_optimiser

GenerativeModelProtocols.earth_mover_distance

GenerativeModelProtocols.earth_mover_distance

GenerativeModelProtocols.earth_mover_distance

GenerativeModelProtocols.encode

GenerativeModelProtocols.encode

GenerativeModelProtocols.energy_distance

GenerativeModelProtocols.energy_distance

GenerativeModelProtocols.energy_distance

GenerativeModelProtocols.input_size

GenerativeModelProtocols.latent_size

GenerativeModelProtocols.sinkhorn_distance

GenerativeModelProtocols.sinkhorn_distance

GenerativeModelProtocols.sinkhorn_distance

GenerativeModelProtocols.train!

Types

GenerativeModelProtocols.GenerativeModelProtocol — Type

struct GenerativeModelProtocol{M<:GenerativeModelProtocols.AbstractGenerativeModel}

A protocol containing a user-specified generative model, and all of its parameters needed to train it and evaluate it. By default the training and evaluation of the generative model is done in the CPU, however, this can be changed by setting device to a GPU device of preference.

Structure Fields

training_data::Matrix{AbstractFloat}: Vector containing all the data, scaled and shifted to have zero mean and unit variance, that is to be used for training.

mean_training_data::Tuple{Vararg{AbstractFloat}}: Mean of the raw (unshifted and unscaled) training data.

var_training_data::Tuple{Vararg{AbstractFloat}}: Variance of the raw (unshifted and unscaled) training data.

normalize_data::Bool: Determines if the training data should be normalized to have zero mean and unit variance when passed to the generative model.

model::GenerativeModelProtocols.AbstractGenerativeModel: Generative model architecture to be used.

_log::Dict{String, Vector}: Stores the time-series data generated while training the generative model.

GenerativeModelProtocols.VariationalAutoencoder — Type

struct VariationalAutoencoder{EncLayerNames, DecLayerNames} <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Variational Autoencoder (VAE). Once trained, its decoder can be used as a generative model.

Structure Fields

encoders::NamedTuple{EncLayerNames, <:Tuple{Vararg{Chain}}} where EncLayerNames: Encoder chain of the VAE, responsible for encoding time-series data into a latent representation.

decoders::NamedTuple{DecLayerNames, <:Tuple{Vararg{Chain}}} where DecLayerNames: Decoder chain of the VAE, responsible for generating time-series data from the latent representation.

_ps::Ref{<:NamedTuple}: Trained parameters of the model. Users should not use this directly.

_st::Ref{<:NamedTuple}: Trained state of the model. Users should not use this directly.

GenerativeModelProtocols.DiffusionModel — Type

struct DiffusionModel{LayerNames} <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Diffusion Model (DM). Once trained, its denoiser model can be used as a generative model.

Structure Fields

β::Tuple{Vararg{AbstractFloat}}: Variance of the Gaussian noise added in each diffusion step.

denoiser_model::GenerativeModelProtocols.TabularDenoiser: Neural network parametrizing the denoising model, that is, a model that probabilisitically undoes the Gaussian noise.

GenerativeModelProtocols.GaussianMixtureModel — Type

struct GaussianMixtureModel{F<:AbstractFloat} <: GenerativeModelProtocols.AbstractCategoricalGenerativeModel

A structure containing the general parameters needed to evaluate and train a Gaussian Mixture Model (GMM). Once trained, it can be used a generative model.

Structure Fields

log_σ²::Matrix{F} where F<:AbstractFloat: Logarithm of the square of the variance of the Gaussian clusters of the model.

μ::Matrix{F} where F<:AbstractFloat: Mean values of the Gaussian clusters of the model.

predictor_network::Chain: Neural network parametrizing the likelihood of some input stemming from a given cluster.

p::Vector{F} where F<:AbstractFloat: Vector containing the categorical probabilities of each cluster.

_ps::Ref{<:NamedTuple}: Trained parameters of the predictor model. Users should not use directly use this.

_st::Ref{<:NamedTuple}: Trained state of the predictor model. Users should not use directly use this.

GenerativeModelProtocols.TabularDenoiser — Type

struct TabularDenoiser{LayerNames}

A time-conditioned residual neural network for vector-based diffusion models.

Structure Fields

T::Int64: Total number of discrete time steps in the forward noising process and reverse denoising timeline.

time_embedding_mlp::Chain: Multi-layer perceptron that maps static sinusoidal time frequencies into a globally shared learned temporal context vector.

input_projection::Dense: Entry layer that projects the raw noisy input data vector into the initial hidden feature state.

residual_layers::NamedTuple{LayerNames, <:Tuple{Vararg{Dense}}} where LayerNames: Tuple of sequential dense layers that transform hidden features inside the recurrent residual block loop.

time_projection_layers::NamedTuple{LayerNames, <:Tuple{Vararg{Dense}}} where LayerNames: Tuple of dense layers that map the shared temporal context into a dynamic additive bias shift inside the recurrent residual block loop.

output_projection::Dense: Exit layer that projects final hidden features out of the residual block loop back to the original data dimensions to output the noise prediction.

max_period::AbstractFloat: Constant that sets the maximum periodic scale for the base sinusoidal time step calculation.

_ps::Ref{<:NamedTuple}: Trained parameters of the tabular denoiser. Users should not use directly use this.

_st::Ref{<:NamedTuple}: Trained state of the tabular denoiser. Users should not use directly use this.

GenerativeModelProtocols.GenerativeAdversarialNetwork — Type

struct GenerativeAdversarialNetwork <: GenerativeModelProtocols.AbstractCategoricalGenerativeModel

A structure containing the general parameters needed to evaluate and train a Generative Adversarial Network (GAN). Once trained, it can be used a generative model.

Structure Fields

discriminator::Chain: Neural network discriminating between real and synthetic data.

vae_model::GenerativeModelProtocols.VariationalAutoencoder: Variational Autoencoder for encoding and decoding.

_ps_discriminator::Ref{<:NamedTuple}: Trained parameters of the discriminator model. Users should not use directly use this.

_st_discriminator::Ref{<:NamedTuple}: Trained state of the discriminator model. Users should not use directly use this.

GenerativeModelProtocols.NormalizingFlow — Type

struct NormalizingFlow <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Continuous Normalizing Flow (CNF). Once trained, it can be used a generative model.

Structure Fields

velocity_field::Chain: Neural network parametrizing the velocity field as a function of time.

_ps::Ref{NamedTuple}: Trained parameters of the model. Users should not use this directly.

_st::Ref{NamedTuple}: Trained state of the model. Users should not use this directly.

Methods

GenerativeModelProtocols.train! — Method

train!(protocol::GenerativeModelProtocol; kwargs...)

Method to train the model inside protocol with its training_data. This method can be called again after it finishes to resume training.

Arguments

protocol: The generative model protocol to train.

Keyword Arguments (All Models)

print_log::Bool: Whether to print training logs (default is true).

epochs::Int: The number of epochs to train for (default is 100).

print_epochs::Int: The number of epochs between each print of the training log (default is 5).

batchsize::Int: The batch size for training (default is 32).

shuffle::Bool: Whether to shuffle the training data (default is true).

optimiser::Optimisers.AbstractRule: The optimiser to use for training (default is GenerativeModelProtocols.default_optimiser(protocol)).

device::Lux.MLDataDevices.AbstractDevice: The device to use for training (default is cpu_device()).

precision::Function: The precision function for training (default is f32).

Keyword Arguments (VariationalAutoencoder)

β::AbstractFloat: The β parameter for the VAE ELBO function (default is 1.0).

Keyword Arguments (GenerativeAdversarialNetwork)

Attention: Unlike all other generative models, GenerativeAdversarialNetwork needs a 2-element tuple of optimisers.

β::AbstractFloat: The β parameter for the VAE ELBO function (default is 1.0).

n_critic::Int: The number of subiterations to train the critic in each training step (default is 5).

γ_vae::AbstractFloat: The weight of the VAE ELBO in the VAE loss function (default is 1.0).

γ_wgan::AbstractFloat: The weight of the Wasserstein distance in the VAE loss function (default is 1.0).

grad_penalty::Bool: Whether to use gradient penalty in the WGAN loss function (default is false).

λ::AbstractFloat: The weight of the gradient penalty in the WGAN loss function (default is 10.0).

a::AbstractFloat: The gradient norm target in the gradient penalty (default is 1.0).

weight_clipping::Bool: Whether to use weight clipping in the WGAN loss function (default is false).

clip_value::AbstractFloat: The value to clip the weights and biases of the critic at each training step (default is 1.0).

optimiser::Tuple{<:Optimisers.AbstractRule, <:Optimisers.AbstractRule}: The optimisers for the generator and critic for training.

GenerativeModelProtocols.categorize — Method

categorize(protocol::GenerativeModelProtocol, x::Vector) -> Vector

Compute the posterior probability distribution over the mixture components for a given input vector x. Returns a vector where the k-th element represents the conditional probability that the input stems from the k-th categorical cluster of the model.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use for categorization.

x::Vector: The input vector for which to compute the posterior probability distribution.

GenerativeModelProtocols.categorize — Method

categorize(protocol::GenerativeModelProtocol, x::Matrix) -> Matrix

Batch compute the posterior probability distributions over the mixture components for multiple input vectors. Each sample in the input matrix x is mapped to a normalized categorical probability vector where the k-th element represents the conditional probability that the sample stems from the k-th cluster.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use for categorization.

x::Matrix: The input matrix for which to compute the posterior probability distributions.

GenerativeModelProtocols.encode — Method

encode(protocol::GenerativeModelProtocol, x::Vector; kwargs...) -> Vector

Encodes the data space variable x into a latent space variable.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use for encoding.

x::Vector: The input vector to encode.

Keyword Arguments (NormalizingFlow)

ode_solver: The ODE solver to use for the normalizing flow.

t_final::AbstractFloat: The final time for the ODE solver.

GenerativeModelProtocols.encode — Method

encode(protocol::GenerativeModelProtocol, x::Matrix; kwargs...) -> Matrix

Encodes the data space variable x into a latent space variable.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use for encoding.

x::Matrix: The input matrix to encode.

Keyword Arguments (NormalizingFlow)

ode_solver: The ODE solver to use for the normalizing flow.

t_final::AbstractFloat: The final time for the ODE solver.

GenerativeModelProtocols.decode — Method

decode(protocol::GenerativeModelProtocol, z::Vector; kwargs...) -> Vector

Decodes the latent space representations z back into the data space.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use for decoding.

z::Vector: The latent space representations to decode.

Keyword Arguments (NormalizingFlow)

ode_solver: The ODE solver to use for the normalizing flow.

t_final::AbstractFloat: The final time for the ODE solver.

GenerativeModelProtocols.decode — Method

decode(protocol::GenerativeModelProtocol, z::Matrix; kwargs...) -> Matrix

Decodes the latent space representations z back into the data space.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use for decoding.

z::Matrix: The latent space representations to decode.

Keyword Arguments (NormalizingFlow)

ode_solver: The ODE solver to use for the normalizing flow.

t_final::AbstractFloat: The final time for the ODE solver.

GenerativeModelProtocols.input_size — Method

input_size(protocol::GenerativeModelProtocol) -> Int

Returns the dimension size of an input vector.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use.

GenerativeModelProtocols.latent_size — Method

latent_size(protocol::GenerativeModelProtocol) -> Int

Returns the dimension size of an encoded latent vector.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use.

GenerativeModelProtocols.default_optimiser — Method

GenerativeModelProtocols.default_optimiser(protocol::GenerativeModelProtocol)

Return the default optimiser for the generative model in the protocol.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol.

GenerativeModelProtocols.sinkhorn_distance — Method

GenerativeModelProtocols.sinkhorn_distance(
    synthetic_data::AbstractMatrix, target_data::AbstractMatrix; 
    metric = Distances.Euclidean(), ε::Float64 = 0.05, 
    alg = SinkhornGibbs(), atol::Float64 = 0.0, 
    rtol::Float64 = atol > 0 ? 0 : √eps(Float64), 
    check_convergence::Int = 10, maxiter::Int = 100_000,
    regularization::Bool = false)

Calculates the Sinkhorn distance between synthetic and target data.

Arguments

synthetic_data::AbstractMatrix: A matrix of synthetic data points

target_data::AbstractMatrix: A matrix of target data points

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

ε::Float64: The regularization parameter for Sinkhorn divergence (default is 0.05)

alg: The algorithm to use for solving the Sinkhorn problem (default is SinkhornGibbs)

atol::Float64: Absolute tolerance for convergence (default is 0)

rtol::Float64: Relative tolerance for convergence (default is 0 if atol > 0, otherwise √eps(Float64))

check_convergence::Int: Number of iterations between convergence checks (default is 10)

maxiter::Int: Maximum number of iterations (default is 100_000)

regularization::Bool: Whether to add the regularization term to the returned distance (default is false)

GenerativeModelProtocols.sinkhorn_distance — Method

GenerativeModelProtocols.sinkhorn_distance(
    protocol::GenerativeModelProtocol, target_data::AbstractMatrix; 
    metric = Distances.Euclidean(), ε::Float64 = 0.05, 
    alg = SinkhornGibbs(), atol::Float64 = 0.0, 
    rtol::Float64 = atol > 0 ? 0 : √eps(Float64), 
    check_convergence::Int = 10, maxiter::Int = 100_000,
    regularization::Bool = false)

Calculates the Sinkhorn distance between synthetic data produced by the protocol and target data.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use

target_data::AbstractMatrix: A matrix of target data points

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

ε::Float64: The regularization parameter for Sinkhorn divergence (default is 0.05)

alg: The algorithm to use for solving the Sinkhorn problem (default is SinkhornGibbs)

atol::Float64: Absolute tolerance for convergence (default is 0)

rtol::Float64: Relative tolerance for convergence (default is 0 if atol > 0, otherwise √eps(Float64))

check_convergence::Int: Number of iterations between convergence checks (default is 10)

maxiter::Int: Maximum number of iterations (default is 100_000)

regularization::Bool: Whether to add the regularization term to the returned distance (default is false)

GenerativeModelProtocols.sinkhorn_distance — Method

GenerativeModelProtocols.sinkhorn_distance(
    protocol::GenerativeModelProtocol; 
    metric = Distances.Euclidean(), ε::Float64 = 0.05, 
    alg = SinkhornGibbs(), atol::Float64 = 0.0, 
    rtol::Float64 = atol > 0 ? 0 : √eps(Float64), 
    check_convergence::Int = 10, maxiter::Int = 100_000,
    regularization::Bool = false)

Calculates the Sinkhorn distance between synthetic data produced by the protocol and its training data.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

ε::Float64: The regularization parameter for Sinkhorn divergence (default is 0.05)

alg: The algorithm to use for solving the Sinkhorn problem (default is SinkhornGibbs)

atol::Float64: Absolute tolerance for convergence (default is 0)

rtol::Float64: Relative tolerance for convergence (default is 0 if atol > 0, otherwise √eps(Float64))

check_convergence::Int: Number of iterations between convergence checks (default is 10)

maxiter::Int: Maximum number of iterations (default is 100_000)

regularization::Bool: Whether to add the regularization term to the returned distance (default is false)

GenerativeModelProtocols.earth_mover_distance — Method

GenerativeModelProtocols.earth_mover_distance(
    synthetic_data::AbstractMatrix, target_data::AbstractMatrix; 
    metric = Distances.Euclidean(),max_iter::Int = 5000)

Calculates the Earth Mover's Distance between synthetic data and target data.

Arguments

synthetic_data::AbstractMatrix: A matrix of synthetic data points

target_data::AbstractMatrix: A matrix of target data points

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

max_iter::Int: Maximum number of iterations for the LP optimizer (default is 5000)

GenerativeModelProtocols.earth_mover_distance — Method

GenerativeModelProtocols.earth_mover_distance(
    protocol::GenerativeModelProtocol, target_data::AbstractMatrix; 
    metric = Distances.Euclidean(), max_iter::Int = 5000)

Calculates the Earth Mover's Distance between synthetic data produced by the protocol and target data.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use

target_data::AbstractMatrix: A matrix of target data points

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

max_iter::Int: Maximum number of iterations for the LP optimizer (default is 5000)

GenerativeModelProtocols.earth_mover_distance — Method

GenerativeModelProtocols.earth_mover_distance(
    protocol::GenerativeModelProtocol; 
    metric = Distances.Euclidean(), max_iter::Int = 5000)

Calculates the Earth Mover's Distance between synthetic data produced by the protocol and its training data.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

max_iter::Int: Maximum number of iterations for the LP optimizer (default is 5000)

GenerativeModelProtocols.energy_distance — Method

GenerativeModelProtocols.energy_distance(
    synthetic_data::AbstractMatrix, target_data::AbstractMatrix;
    metric = Distances.Euclidean())

Calculates the Energy Distance between synthetic and target data.

Arguments

synthetic_data::AbstractMatrix: A matrix of synthetic data points

target_data::AbstractMatrix: A matrix of target data points

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

GenerativeModelProtocols.energy_distance — Method

GenerativeModelProtocols.energy_distance(
    protocol::GenerativeModelProtocol, target_data::AbstractMatrix; 
    metric = Distances.Euclidean())

Calculates the Energy Distance between synthetic data produced by the protocol and target data.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use

target_data::AbstractMatrix: A matrix of target data points

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

GenerativeModelProtocols.energy_distance — Method

GenerativeModelProtocols.energy_distance(
    protocol::GenerativeModelProtocol; 
    metric = Distances.Euclidean())

Calculates the Energy Distance between synthetic data produced by the protocol and its training data.

Arguments

protocol::GenerativeModelProtocol: The generative model protocol to use

Keyword Arguments

metric::Metric: The distance metric to use (default is Euclidean)

Convenience Constructors

GenerativeModelProtocols.GenerativeModelProtocol — Method

GenerativeModelProtocol(model::M, training_data::Matrix{<:AbstractFloat}; 
    normalize_data::Bool = true, kwargs...) where {M<:AbstractGenerativeModel}

Convenience constructor to create a GenerativeModelProtocol with default training parameters, optimiser and compute device.

Arguments

model<:AbstractGenerativeModel: The generative model to use.

training_data::Matrix{<:AbstractFloat}: The training data to use.

Keyword Arguments

normalize_data::Bool = true: Whether to normalize the training data.

kwargs...: Additional keyword arguments for the generative model protocol.

GenerativeModelProtocols.VariationalAutoencoder — Method

GenerativeModelProtocols.VariationalAutoencoder(
    encoders::Tuple{Vararg{Lux.Chain}}, decoders::Tuple{Vararg{Lux.Chain}})

Convenience constructor that generates a GenerativeModelProtocols.VariationalAutoencoder from plain Tuple containers.

Arguments

encoders::Tuple{Vararg{Lux.Chain}}: The encoder networks.

decoders::Tuple{Vararg{Lux.Chain}}: The decoder networks.

GenerativeModelProtocols.VariationalAutoencoder — Method

GenerativeModelProtocols.VariationalAutoencoder(
    encoders::NamedTuple{EncLayerNames, Tuple{Vararg{Lux.Chain}}}, 
    decoders::NamedTuple{DecLayerNames, Tuple{Vararg{Lux.Chain}}}) 
    where {EncLayerNames, DecLayerNames}

Convenience constructor that generates a GenerativeModelProtocols.VariationalAutoencoder from plain Tuple containers.

Arguments

encoders::NamedTuple{EncLayerNames, Tuple{Vararg{Lux.Chain}}}: The encoder networks.

decoders::NamedTuple{DecLayerNames, Tuple{Vararg{Lux.Chain}}}: The decoder networks.

GenerativeModelProtocols.VariationalAutoencoder — Method

GenerativeModelProtocols.VariationalAutoencoder(
    input_dim::Int, latent_dim::Int = 1, latent_layers::Int = 1)

Convenience constructor that generates a GenerativeModelProtocols.VariationalAutoencoder using default encoder and decoder network architectures.

Arguments

input_dim::Int: The dimensionality of the input data.

latent_dim::Int: The dimensionality of the latent space(s) (default is 1).

latent_layers::Int: The number of latent layers (default is 1).

GenerativeModelProtocols.GaussianMixtureModel — Method

GenerativeModelProtocols.GaussianMixtureModel(input_size::Int, k::Int)

Convenience constructor that creates a GenerativeModelProtocols.GaussianMixtureModel using default predictor network architecture.

Arguments

input_size::Int: The dimensionality of the input data.

k::Int: The number of Gaussian clusters in the model.

GenerativeModelProtocols.DiffusionModel — Method

GenerativeModelProtocols.DiffusionModel(num_inputs::Int, β::Tuple{Vararg{AbstractFloat}}})

Convenience constructor to create a GenerativeModelProtocols.DiffusionModel.

A default denoiser_model is created based on num_inputs.

Arguments

num_inputs::Int: The dimensionality of the input data.

β::Tuple{Vararg{AbstractFloat}}: The variance of the Gaussian noise added in each diffusion step.

GenerativeModelProtocols.DiffusionModel — Method

GenerativeModelProtocols.DiffusionModel(
    T::Int, β_start::FP, β_end::FP, denoiser_model::TabularDenoiser{LayerNames}) 
    where {LayerNames, FP<:AbstractFloat}

Convenience constructor to create a GenerativeModelProtocols.DiffusionModel.

This constructor initializes β with collect(range(β_start, β_end, length=T)).

Arguments

T::Int: The number of diffusion steps.

β_start::FP: The variance of the Gaussian noise at the first step.

β_end::FP: The variance of the Gaussian noise at the last step.

denoiser_model::TabularDenoiser{LayerNames}: The neural network parametrizing the denoising model.

GenerativeModelProtocols.DiffusionModel — Method

GenerativeModelProtocols.DiffusionModel(
    num_inputs::Int, T::Int = 5, β_start::FP = 1.0E-4, β_end::FP = 0.02) 
    where {FP<:AbstractFloat}

Convenience constructor to create a GenerativeModelProtocols.DiffusionModel.

A default denoiser_model is created based on num_inputs. This constructor initializes β with collect(range(β_start, β_end, length=T)).

Arguments

num_inputs::Int: The dimensionality of the input data.

T::Int: The number of diffusion steps.

β_start::FP: The variance of the Gaussian noise at the first step.

β_end::FP: The variance of the Gaussian noise at the last step.

GenerativeModelProtocols.TabularDenoiser — Method

GenerativeModelProtocols.TabularDenoiser(num_inputs::Int; 
    hidden_layer_size::Int = 64, T::Int = 32, activation_function::Function = relu, 
    max_period::Float64 = 10000.0)

Convenience constructor to create a default TabularDenoiser that matches the number of inputs specified by num_inputs.

Arguments

num_inputs::Int: The dimensionality of the input data.

Keyword Arguments

hidden_layer_size::Int: The size of the hidden layers.

T::Int: The number of time steps.

activation_function::Function: The activation function to use.

max_period::Float64: The maximum period for the time embedding.

GenerativeModelProtocols.GenerativeAdversarialNetwork — Method

GenerativeModelProtocols.GenerativeAdversarialNetwork(
    input_size::Int, latent_size::Int = 1, latent_layers::Int = 1)

Convenience constructor that creates a GenerativeModelProtocols.GenerativeAdversarialNetwork using default generator and discriminator networks.

Arguments

input_size::Int: The dimensionality of the input data.

latent_size::Int: The size of the latent space (default is 1).

latent_layers::Int: The number of layers in the latent space (default is 1).

GenerativeModelProtocols.NormalizingFlow — Method

GenerativeModelProtocols.NormalizingFlow(input_size::Int)

Convenience constructor that creates a GenerativeModelProtocols.NormalizingFlow using default velocity field network architecture.

Arguments

input_size::Int: The dimensionality of the input data.