Reference

Contents

Reference

Contents

Index

Index

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.GenerativeModelProtocol

GenerativeModelProtocols.VariationalAutoencoder

GenerativeModelProtocols.DiffusionModel — Type

struct DiffusionModel <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Diffusion Model (DM). Once trained, its denoiser model can be used as a generative model.

T::Int64: Number of diffusion steps (number of steps in the Markov chain).

α::Vector{Float64}: Float64(1.0) .- β

ᾱ::Vector{Float64}: cumprod(α)

β::Vector{Float64}: Variance of the Gaussian noise added in each diffusion step.

denoiser_model::Flux.Chain: Neural network parametrizing the denoising model, that is, a model that probabilisitically undoes the Gaussian noise.

GenerativeModelProtocols.GenerativeModelProtocol — Type

struct GenerativeModelProtocol{M<:GenerativeModelProtocols.AbstractGenerativeModel}

A protocol containing a user-specified generative model, and all of its parameters needed to train it and evaluate it. By default the training and evaluation of the generative model is done in the CPU, however, this can be changed by setting device to a GPU device of preference.

training_data::Matrix{Float64}: Vector containing all the data that is to be used for training.

epochs::Int64: Number of training epochs for the generative model.

batchsize::Int64: Batch size for training the generative model.

shuffle::Bool: Flag indicating whether to shuffle the training data during training.

optimiser::Union{Flux.Optimise.AbstractOptimiser, Optimisers.AbstractRule}: Optimizer used to train the generative model.

device::MLDataDevices.AbstractDevice: Hardware device (CPU or GPU) on which to perform training and inference.

model::GenerativeModelProtocols.AbstractGenerativeModel: Generative model architecture to be used.

_log::GenerativeModelProtocols.TrainingLog: Stores the time-series data generated while training the generative model.

GenerativeModelProtocols.VariationalAutoencoder — Type

struct VariationalAutoencoder <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Variational Autoencoder (VAE). Once trained, its decoder can be used as a generative model.

latent_dim::Int64: Dimensionality of the latent space representation.

latent_layers::Int64: Number of latent layers in the encoder and decoder chains of the VAE.

encoders::Tuple{Vararg{Flux.Chain}}: Encoder chain of the VAE, responsible for encoding time-series data into a latent representation.

decoders::Tuple{Vararg{Flux.Chain}}: Decoder chain of the VAE, responsible for generating time-series data from the latent representation.

β::Union{Float64, Vector{Float64}}: Weighting factor for the KL divergence term in the VAE loss function.