Reference

Contents

Reference

Contents

Index

Types

Methods

Convenience Constructors

Index

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.DiffusionModel

GenerativeModelProtocols.GaussianMixtureModel

GenerativeModelProtocols.GaussianMixtureModel

GenerativeModelProtocols.GenerativeAdversarialNetwork

GenerativeModelProtocols.GenerativeAdversarialNetwork

GenerativeModelProtocols.GenerativeModelProtocol

GenerativeModelProtocols.GenerativeModelProtocol

GenerativeModelProtocols.NormalizingFlow

GenerativeModelProtocols.NormalizingFlow

GenerativeModelProtocols.TabularDenoiser

GenerativeModelProtocols.TabularDenoiser

GenerativeModelProtocols.VariationalAutoencoder

GenerativeModelProtocols.VariationalAutoencoder

GenerativeModelProtocols.categorize

GenerativeModelProtocols.categorize

GenerativeModelProtocols.decode

GenerativeModelProtocols.decode

GenerativeModelProtocols.encode

GenerativeModelProtocols.encode

GenerativeModelProtocols.input_size

GenerativeModelProtocols.latent_size

GenerativeModelProtocols.train!

Types

GenerativeModelProtocols.GenerativeModelProtocol — Type

struct GenerativeModelProtocol{M<:GenerativeModelProtocols.AbstractGenerativeModel}

A protocol containing a user-specified generative model, and all of its parameters needed to train it and evaluate it. By default the training and evaluation of the generative model is done in the CPU, however, this can be changed by setting device to a GPU device of preference.

training_data::Union{Nothing, Matrix}: Vector containing all the data, scaled and shifted to have zero mean and unit variance, that is to be used for training.

mean_training_data::Tuple: Mean of the raw (unshifted and unscaled) training data.

var_training_data::Tuple: Variance of the raw (unshifted and unscaled) training data.

epochs::Int64: Number of training epochs for the generative model.

batchsize::Int64: Batch size for training the generative model.

shuffle::Bool: Flag indicating whether to shuffle the training data during training.

normalize_data::Bool: Determines if the training data should be normalized to have zero mean and unit variance when passed to the generative model.

optimiser::Any: Optimizer used to train the generative model.

device::MLDataDevices.AbstractDevice: Hardware device (CPU or GPU) on which to perform training and inference.

model::GenerativeModelProtocols.AbstractGenerativeModel: Generative model architecture to be used.

precision::Function: Floating point precision to be used

_log::Dict{String, Vector}: Stores the time-series data generated while training the generative model.

GenerativeModelProtocols.VariationalAutoencoder — Type

struct VariationalAutoencoder{EncLayerNames, DecLayerNames} <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Variational Autoencoder (VAE). Once trained, its decoder can be used as a generative model.

encoders::NamedTuple{EncLayerNames, <:Tuple{Vararg{Chain}}} where EncLayerNames: Encoder chain of the VAE, responsible for encoding time-series data into a latent representation.

decoders::NamedTuple{DecLayerNames, <:Tuple{Vararg{Chain}}} where DecLayerNames: Decoder chain of the VAE, responsible for generating time-series data from the latent representation.

_ps::Union{Nothing, Ref{<:NamedTuple}}: Trained parameters of the model. Users should not use this directly.

_st::Union{Nothing, Ref{<:NamedTuple}}: Trained state of the model. Users should not use this directly.

GenerativeModelProtocols.DiffusionModel — Type

struct DiffusionModel{LayerNames, F<:AbstractFloat} <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Diffusion Model (DM). Once trained, its denoiser model can be used as a generative model.

β::Tuple{Vararg{F}} where F<:AbstractFloat: Variance of the Gaussian noise added in each diffusion step.

denoiser_model::GenerativeModelProtocols.TabularDenoiser: Neural network parametrizing the denoising model, that is, a model that probabilisitically undoes the Gaussian noise.

GenerativeModelProtocols.GaussianMixtureModel — Type

struct GaussianMixtureModel{F<:AbstractFloat} <: GenerativeModelProtocols.AbstractCategoricalGenerativeModel

A structure containing the general parameters needed to evaluate and train a Gaussian Mixture Model (GMM). Once trained, it can be used a generative model.

log_σ²::Matrix{F} where F<:AbstractFloat: Logarithm of the square of the variance of the Gaussian clusters of the model.

μ::Matrix{F} where F<:AbstractFloat: Mean values of the Gaussian clusters of the model.

predictor_network::Chain: Neural network parametrizing the likelihood of some input stemming from a given cluster.

p::Vector{F} where F<:AbstractFloat: Vector containing the categorical probabilities of each cluster.

_ps::Union{Nothing, Ref{<:NamedTuple}}: Trained parameters of the predictor model. Users should not use directly use this.

_st::Union{Nothing, Ref{<:NamedTuple}}: Trained state of the predictor model. Users should not use directly use this.

GenerativeModelProtocols.TabularDenoiser — Type

struct TabularDenoiser{LayerNames, F<:AbstractFloat}

A time-conditioned residual neural network for vector-based diffusion models.latentdim, latentlayers, encoders, decoders

T::Int64: Total number of discrete time steps in the forward noising process and reverse denoising timeline.

time_embedding_mlp::Chain: Multi-layer perceptron that maps static sinusoidal time frequencies into a globally shared learned temporal context vector.

input_projection::Dense: Entry layer that projects the raw noisy input data vector into the initial hidden feature state.

residual_layers::NamedTuple{LayerNames, <:Tuple{Vararg{Dense}}} where LayerNames: Tuple of sequential dense layers that transform hidden features inside the recurrent residual block loop.

time_projection_layers::NamedTuple{LayerNames, <:Tuple{Vararg{Dense}}} where LayerNames: Tuple of dense layers that map the shared temporal context into a dynamic additive bias shift inside the recurrent residual block loop.

output_projection::Dense: Exit layer that projects final hidden features out of the residual block loop back to the original data dimensions to output the noise prediction.

max_period::AbstractFloat: Constant that sets the maximum periodic scale for the base sinusoidal time step calculation.

_ps::Union{Nothing, Ref{<:NamedTuple}}: Trained parameters of the tabular denoiser. Users should not use directly use this.

_st::Union{Nothing, Ref{<:NamedTuple}}: Trained state of the tabular denoiser. Users should not use directly use this.

GenerativeModelProtocols.GenerativeAdversarialNetwork — Type

struct GenerativeAdversarialNetwork <: GenerativeModelProtocols.AbstractCategoricalGenerativeModel

A structure containing the general parameters needed to evaluate and train a Generative Adversarial Network (GAN). Once trained, it can be used a generative model.

discriminator::Chain: Neural network discriminating between real and synthetic data.

vae_model::GenerativeModelProtocols.VariationalAutoencoder: Variational Autoencoder for encoding and decoding.

_ps_discriminator::Union{Nothing, Ref{<:NamedTuple}}: Trained parameters of the discriminator model. Users should not use directly use this.

_st_discriminator::Union{Nothing, Ref{<:NamedTuple}}: Trained state of the discriminator model. Users should not use directly use this.

GenerativeModelProtocols.NormalizingFlow — Type

struct NormalizingFlow <: GenerativeModelProtocols.AbstractGenerativeModel

A structure containing the general parameters needed to evaluate and train a Continuous Normalizing Flow (CNF). Once trained, it can be used a generative model.

velocity_field::Chain: Neural network parametrizing the velocity field as a function of time.

_ps::Union{Nothing, Ref{<:NamedTuple}}: Trained parameters of the model. Users should not use this directly.

_st::Union{Nothing, Ref{<:NamedTuple}}: Trained state of the model. Users should not use this directly.

Methods

GenerativeModelProtocols.train! — Method

train!(protocol::GenerativeModelProtocol; kwargs...)

Method to train the model inside protocol with its training_data. This method can be called again after it finishes to resume training.

GenerativeModelProtocols.categorize — Method

categorize(protocol::GenerativeModelProtocol, x::Vector) -> Vector

Compute the posterior probability distribution over the mixture components for a given input vector x. Returns a vector where the k-th element represents the conditional probability that the input stems from the k-th categorical cluster of the model.

GenerativeModelProtocols.categorize — Method

categorize(protocol::GenerativeModelProtocol, x::Matrix) -> Matrix

Batch compute the posterior probability distributions over the mixture components for multiple input vectors. Each sample in the input matrix x is mapped to a normalized categorical probability vector where the k-th element represents the conditional probability that the sample stems from the k-th cluster.

GenerativeModelProtocols.encode — Method

encode(protocol::GenerativeModelProtocol, x::Vector; kwargs...) -> Vector

Encodes the data space variable x into a latent space variable.

GenerativeModelProtocols.encode — Method

encode(protocol::GenerativeModelProtocol, x::Matrix; kwargs...) -> Matrix

Encodes the data space variable x into a latent space variable.

GenerativeModelProtocols.decode — Method

decode(protocol::GenerativeModelProtocol, z::Vector; kwargs...) -> Vector

Decodes the latent space representations z back into the data space.

GenerativeModelProtocols.decode — Method

decode(protocol::GenerativeModelProtocol, z::Matrix; kwargs...) -> Matrix

Decodes the latent space representations z back into the data space.

GenerativeModelProtocols.input_size — Method

input_size(protocol::GenerativeModelProtocol) -> Int

Returns the dimension size of an input vector.

GenerativeModelProtocols.latent_size — Method

latent_size(protocol::GenerativeModelProtocol) -> Int

Returns the dimension size of an encoded latent vector.

Convenience Constructors

GenerativeModelProtocols.GenerativeModelProtocol — Method

GenerativeModelProtocol(model::M, training_data::Matrix{<:AbstractFloat}; normalize_data::Bool = true, kwargs...) where {M<:AbstractGenerativeModel}

Convenience constructor to create a GenerativeModelProtocol with default training parameters, optimiser and compute device.

GenerativeModelProtocols.VariationalAutoencoder — Method

GenerativeModelProtocols.VariationalAutoencoder(input_dim::Int, latent_dim::Int = 1, latent_layers::Int = 1)

Convenience constructor that generates a GenerativeModelProtocols.VariationalAutoencoder using default encoder and decoder network architectures.

GenerativeModelProtocols.GaussianMixtureModel — Method

GenerativeModelProtocols.GaussianMixtureModel(input_size::Int, k::Int)

Convenience constructor that creates a GenerativeModelProtocols.GaussianMixtureModel using default predictor network architecture.

GenerativeModelProtocols.DiffusionModel — Method

GenerativeModelProtocols.DiffusionModel(num_inputs::Int, β::Tuple{Vararg{AbstractFloat}}})

Convenience constructor to create a GenerativeModelProtocols.DiffusionModel.

A default denoiser_model is created based on num_inputs.

GenerativeModelProtocols.DiffusionModel — Method

GenerativeModelProtocols.DiffusionModel(T::Int, β_start::FP, β_end::FP, denoiser_model::TabularDenoiser{LayerNames, FP}) where {LayerNames, FP<:AbstractFloat}

Convenience constructor to create a GenerativeModelProtocols.DiffusionModel.

This constructor initializes β with collect(range(β_start, β_end, length=T)).

GenerativeModelProtocols.DiffusionModel — Method

GenerativeModelProtocols.DiffusionModel(num_inputs::Int, T::Int = 5, β_start::FP = 1.0E-4, β_end::FP = 0.02) where {FP<:AbstractFloat}

Convenience constructor to create a GenerativeModelProtocols.DiffusionModel.

A default denoiser_model is created based on num_inputs. This constructor initializes β with collect(range(β_start, β_end, length=T)).

GenerativeModelProtocols.TabularDenoiser — Method

GenerativeModelProtocols.TabularDenoiser(num_inputs::Int; hidden_layer_size::Int = 64, T::Int = 32, activation_function::Function = relu, max_period::Float64 = 10000.0)

Convenience constructor to create a default TabularDenoiser that matches the number of inputs specified by num_inputs.

GenerativeModelProtocols.GenerativeAdversarialNetwork — Method

GenerativeModelProtocols.GenerativeAdversarialNetwork(input_size::Int, latent_size::Int)

Convenience constructor that creates a GenerativeModelProtocols.GenerativeAdversarialNetwork using default generator and discriminator networks.

GenerativeModelProtocols.NormalizingFlow — Method

GenerativeModelProtocols.NormalizingFlow(input_size::Int)

Convenience constructor that creates a GenerativeModelProtocols.NormalizingFlow using default velocity field network architecture.